🚿 cleaned the code up a lil bit, ADDED reinterpernt_cast<>()'s and also changed most of the "DWORD's" to "uintptr_t's".
also i changed the bhop code a lil bit, too when you activate it, it will not hop if you have previously hopped.

otherwise the code looks cleaner.

// add a way to close the glow window.

//  make a global colour for "GLOW".

+++ newhack/ModuleGet.h
@@ -4,11 +4,11 @@
 class ModuleGet
 {
 public:
-	DWORD getCLIENT_DLL();
-	DWORD getLocalPlayer();
+	uintptr_t getCLIENT_DLL();
+	uintptr_t getLocalPlayer();
 	int getplayerHealth();
-	DWORD getEntityList();
-	DWORD getGlowObjectManager();
+	uintptr_t getEntityList();
+	uintptr_t getGlowObjectManager();
 	int getLocalTeam();
 	int getCurrentEntityHealth(DWORD currentEntity);
 	int getCurrentEntityDormant(DWORD currentEntity);
